pyproject.toml
[project]
name = "w3d5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.5.0",
    "pymysql>=1.1.2",
    "uvicorn>=0.40.0",
]
